Variable Fortmat Change (#1)
* Fixed device specific metavar handling with no default metavar value

* Updated test JSON

* Changed FMG active detection

* Fixed error handling

* Simplified the API calls by wrapping the UA in an anonymous subroutine that contains the FMG URI, SID, and insecure flag. Modified all the calling functions to accept this

* Switched to URI module in API call sub, setting the path and removing any queries

* URI module

URI module is now installed in the dynamic and deployment phases of the actions

* Updated FMG wait time

* Changed Variables JSON and JSON Schema

Variables are now K:V pairs with the name and the value. As well, the values can be structured and are then flattened into vars

* Changed variable structure

Variables can now have structure and are flattened.

As well, we now only configure in the ADOM and global device. VDOMs TBA

* Updated README with flattened variables

* Actions run only on master branch

* JSON scheme variable types